reorganized functions, simplified operations
