+++ src/arguments.c
@@ -5,8 +5,8 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: mpuig-ma <mpuig-ma@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/03/16 15:52:36 by mpuig-ma          #+#    #+#             */
-/*   Updated: 2023/03/29 17:00:13 by mpuig-ma         ###   ########.fr       */
+/*   Created: 2023/04/02 10:52:24 by mpuig-ma          #+#    #+#             */
+/*   Updated: 2023/04/02 10:59:44 by mpuig-ma         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
