--- src/arguments_bonus.c
@@ -1,1 +1,0 @@
-arguments.c
