--- src/exit.c
@@ -1,19 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   exit.c                                             :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mpuig-ma <mpuig-ma@student.42barcel>       +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/03/16 22:20:17 by mpuig-ma          #+#    #+#             */
-/*   Updated: 2023/03/21 11:49:03 by mpuig-ma         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "push_swap.h"
-
-void	ft_exit(int n)
-{
-	ft_putstr_fd("Error\n", STDERR_FILENO);
-	exit (n);
-}
