--- src/exit_bonus.c
@@ -1,1 +1,0 @@
-exit.c
