+++ src/ft_issorted.c
@@ -5,8 +5,8 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: mpuig-ma <mpuig-ma@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/03/21 15:05:19 by mpuig-ma          #+#    #+#             */
-/*   Updated: 2023/03/30 12:31:57 by mpuig-ma         ###   ########.fr       */
+/*   Created: 2023/04/02 12:02:00 by mpuig-ma          #+#    #+#             */
+/*   Updated: 2023/04/02 12:02:17 by mpuig-ma         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
