--- src/lists_other_bonus.c
@@ -1,1 +1,0 @@
-lists_other.c
