+++ src/logic_mini.c
@@ -5,8 +5,8 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: mpuig-ma <mpuig-ma@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/03/28 15:07:24 by mpuig-ma          #+#    #+#             */
-/*   Updated: 2023/03/30 21:42:49 by mpuig-ma         ###   ########.fr       */
+/*   Created: 2023/04/02 11:12:27 by mpuig-ma          #+#    #+#             */
+/*   Updated: 2023/04/02 14:05:10 by mpuig-ma         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,38 +19,54 @@ int	ft_do_logic_mini(t_data *data)
 	if (data->size == 2)
 	{
 		if (data->a->numbers->content > data->a->numbers->next->content)
-			ft_ra(data);
+			rotate(data->a);
 	}
 	else if (data->size <= 5)
 	{
 		while (ft_lstsize(data->a->numbers) != 3)
-		{
 			ft_pb_smallest(data);
-		}
-		ft_case_three(data);
+		ft_case_three(data->a);
 		while (data->b->numbers != NULL)
-			ft_pa(data);
+			push(data->b, data->a);
 	}
 	return (0);
 }
 
-void	ft_case_three(t_data *data)
+int	ft_case_three(t_stack *stack)
 {
 	int	xyz[3];
 
 	xyz[0] = 0;
 	xyz[1] = 0;
 	xyz[2] = 0;
-	if (data->a->numbers->content > data->a->numbers->next->content)
+	if (stack->numbers->content > stack->numbers->next->content)
 		xyz[0] = 1;
-	if (data->a->numbers->next->content > data->a->numbers->next->next->content)
+	if (stack->numbers->next->content > stack->numbers->next->next->content)
 		xyz[1] = 1;
-	if (data->a->numbers->next->next->content > data->a->numbers->content)
+	if (stack->numbers->next->next->content > stack->numbers->content)
 		xyz[2] = 1;
 	if ((xyz[0] + xyz[1] + xyz[2]) >= 2)
-		ft_sa(data);
+		swap(stack);
 	if (xyz[1] == xyz[2])
-		ft_ra(data);
+		rotate(stack);
 	if (xyz[1] == 1 && xyz[2] == 0)
-		ft_rra(data);
+		rev_rotate(stack);
+	return (0);
+}
+
+void	ft_pb_smallest(t_data *data)
+{
+	t_list	*smallest;
+	int		position;
+
+	smallest = ft_find(data->a, SMALLEST);
+	position = ft_lst_position(data->a->numbers, smallest);
+	while (data->a->numbers->content != smallest->content)
+	{
+		if ((data->size - position) >= data->size / 2)
+			rotate(data->a);
+		else
+			rev_rotate(data->a);
+	}
+	push(data->a, data->b);
 }
