--- src/logic_mini_bonus.c
@@ -1,1 +1,0 @@
-logic_mini.c
