+++ src/main.c
@@ -5,12 +5,14 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: mpuig-ma <mpuig-ma@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/03/16 14:56:31 by mpuig-ma          #+#    #+#             */
-/*   Updated: 2023/03/30 12:55:48 by mpuig-ma         ###   ########.fr       */
+/*   Created: 2023/03/31 14:17:17 by mpuig-ma          #+#    #+#             */
+/*   Updated: 2023/04/02 10:59:27 by mpuig-ma         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
+
+static t_data	*ft_init_data(void);
 
 int	main(int argc, char **argv)
 {
@@ -34,4 +36,24 @@ int	main(int argc, char **argv)
 		ft_do_chunk_method(data);
 	ft_free(data);
 	return (0);
+}
+
+static t_data	*ft_init_data(void)
+{
+	t_data	*data;
+
+	data = ft_calloc(1, sizeof(t_data));
+	if (data == NULL)
+		return (NULL);
+	data->a = ft_calloc(1, sizeof(t_stack));
+	data->b = ft_calloc(1, sizeof(t_stack));
+	if (data->a == NULL || data->b == NULL)
+		return (NULL);
+	data->a->id = 'a';
+	data->b->id = 'b';
+	data->fd = STDOUT_FILENO;
+	data->size = 0;
+	data->n_moves = 0;
+	data->chunk_size = 0;
+	return (data);
 }
