--- src/print_bonus.c
@@ -1,1 +1,0 @@
-print.c
